lambda/requirements.txt
ask-sdk-core
ask-sdk-s3-persistence-adapter
boto3
psycopg2-binary
python-dotenv